+++ toys-core/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-### Unreleased
+### 0.7.0 / Unreleased
 
 * ADDED: Flag groups, which enforce policies around which flags are required.
 * CHANGED: Flags within a group are sorted in help screens.